List: - Housekeeping: solved some "advaced" warnings - Ready to test number of cols constraint effectiveness - Update SPH with latest version - Compile warning are meant (for now)

+++ COMMON/LKH-3/INCLUDE/extract_routes.hpp
@@ -37,7 +37,7 @@ public:
         prev = Prev;
     };
 
-    int is_feasible() const { return size >= MTSPMinSize; }
+    int is_feasible() const { return size >= (size_t)MTSPMinSize; }
     inline GainType get_length() const { return current_length; };
     inline size_t get_size() const { return size; };
     inline bool empty() const { return size == 0; }
